requirements.txt
yfinance
python-telegram-bot
black
